requirements.txt
discord~=2.2.2
python-dotenv~=1.0.0
revChatGPT~=3.3.4
